Separated the protocol conformations of CreateOrderViewController in extensions.
